assets: clean up tests and add single/double
